3 changes
added log(score)
removed stop words
if word not in index dont score
length>2 dont add

+++ search/DocScore.java
@@ -3,9 +3,9 @@ package search;
 public class DocScore {
 	
 	public int docId;
-	public int docScore;
+	public float docScore;
 	
-	DocScore(int docId, int docScore)
+	DocScore(int docId, float docScore)
 	{
 		this.docId = docId;
 		this.docScore = docScore;
